update code v2

+++ DrinkToDoor.Web/Pages/Admins/Ingredients/Index.cshtml.cs
@@ -3,6 +3,7 @@ using DrinkToDoor.Business.Dtos.ParamSearchs;
 using DrinkToDoor.Business.Dtos.Requests;
 using DrinkToDoor.Business.Dtos.Responses;
 using DrinkToDoor.Business.Interfaces;
+using DrinkToDoor.Data.Entities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
@@ -20,6 +21,7 @@ namespace DrinkToDoor.Web.Pages.Admins.Ingredients
         public IEnumerable<IngredientResponse> IngredientResponses { get; set; }
 
         public IngredientRequest IngredientRequest { get; set; } = new IngredientRequest();
+        public IngredientResponse IngredientResponse { get; set; } = new IngredientResponse();
         public int TotalPages { get; set; }
         [BindProperty(SupportsGet = true)]
         public int PageSize { get; set; } = 10;
@@ -43,6 +45,15 @@ namespace DrinkToDoor.Web.Pages.Admins.Ingredients
 
         public async Task<IActionResult> OnPostDeleteAsync(Guid id)
         {
+            if (!ModelState.IsValid)
+            {
+                return Page();
+            }
+            var entity = await _ingredientService.GetByIdAsync(id);
+            if (entity == null)
+            {
+                return NotFound();
+            }
             await _ingredientService.DeleteAsync(id);
             return RedirectToPage();
         }
